requirements.txt
customtkinter>=5.2.0
qrcode[pil]>=7.4.2
Pillow>=10.0.0